handleConnection: add horse drug to recover from panics
The whisper-go library needs to do much more checking of validity of the
whisper files, and report errors.  In certain cases the whisper-go
library panics instead of erroring, which caused the whole carbonwriter
to go down (issue #1).  Avoid that from happening by handling the panic.
The real solution would be to make the whisper-go library robust to
allow graceful handling of errors.